moving data pulling functions into use-layout-data
