+++ src/modules/collections/templates/index.tsx
@@ -3,7 +3,7 @@ import getNumberOfSkeletons from "@lib/util/get-number-of-skeletons"
 import repeat from "@lib/util/repeat"
 import ProductPreview from "@modules/products/components/product-preview"
 import SkeletonProductPreview from "@modules/skeletons/components/skeleton-product-preview"
-import { fetchCollectionProducts } from "@pages/collections/[handle]"
+import { fetchCollectionProducts } from "@lib/hooks/use-layout-data"
 import { useInfiniteQuery } from "@tanstack/react-query"
 import React, { useEffect } from "react"
 import { useInView } from "react-intersection-observer"
